+++ crates/cli/tests/plugins_test.rs
@@ -1,5 +1,6 @@
 mod utils;
 
+use proto_core::flow::install::InstallOptions;
 use proto_core::{
     load_tool_from_locator, warpgate::FileLocator, warpgate::UrlLocator, Id, PluginLocator,
     ProtoEnvironment, Tool, UnresolvedVersionSpec,
@@ -26,9 +27,12 @@ where
 
     let mut tool = factory(&proto).await.unwrap();
 
-    tool.setup(&UnresolvedVersionSpec::parse("1.0.0").unwrap(), false)
-        .await
-        .unwrap();
+    tool.setup(
+        &UnresolvedVersionSpec::parse("1.0.0").unwrap(),
+        InstallOptions::default(),
+    )
+    .await
+    .unwrap();
 
     assert!(tool.get_product_dir().exists());